fix(assiduidade): alinha assiduidade com a spec 09/09 (anexos, etapa, despacho)
Lacunas encontradas na auditoria da spec 09/09 contra o codigo. Todas
verificadas live contra a Oracle de dev.

Anexos (chave de escrita != chave de leitura, o GET nunca devolvia nada):
- marcar falta: a leitura passa a usar RH_T_PEDIDO, que e onde a escrita os poe
- ferias: o sync procurava por 'FERIA' e gravava como RH_T_FERIAS_GOZADAS, logo
  nunca encontrava os existentes e duplicava a cada edicao
- hora extra: a validacao procurava por pedido.uuid, mas sao gravados por he.uuid

Validacao:
- ferias: a validacao so procurava TipoAcao.INSERT, mas a alteracao de ferias
  grava UPDATE. A validacao da alteracao ficava pendente para sempre, com o
  pedido ja despachado. Passa a tentar INSERT e cair para UPDATE
- hora extra: a remuneracao era registada antes do recalculo, pelo que
  RH_T_DEF_REMUNERACOES levava o valor pre-ajuste
- falta: grava RH_T_FALTA.DESPACHO_RH = SIM/NAO, como a dispensa e as ferias ja
  fazem em DECISAO_RH. A coluna e VARCHAR2(3) e e para isso mesmo
- falta, dispensa e hora extra: ETAPA fecha em FINALIZADO tambem quando rejeita

Outros:
- FLG_DESCONTO_SAL passa a ser gravado no registo, derivado do tipo de falta:
  uma falta pendente ja diz se desconta salario
- FaltaItemDTO.valorAusencia passa a BigDecimal (Integer truncava os centimos:
  6344 em vez de 6344,56) -- altera o contrato do GET de justificacao

Nao alterado, por decisao: a spec diz DEF_PAGAMENTOS/DEF_PAG_ID na validacao da
justificacao (:857), mas RH_T_FALTA nao tem coluna DEF_PAG_ID -- so DEF_REM_ID,
com FK para RH_T_DEF_REMUNERACOES. E erro do documento, que ficou por actualizar
quando o bloco equivalente do marcar falta foi corrigido a 09/09.

+++ src/main/java/cv/inps/rh/assiduidade/application/dto/FaltaItemDTO.java
@@ -8,6 +8,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -32,7 +34,8 @@ public class FaltaItemDTO  {
   private String horasAusencia ;
 
 
-  private Integer valorAusencia ;
+  /** Valor da falta do dia. BigDecimal: como Integer truncava os cêntimos (6344 em vez de 6344,56). */
+  private BigDecimal valorAusencia ;
 
 
   /** Só de resposta — o motivo grava-se no cabeçalho (aplica-se a todas as seleccionadas). */
